controller tests added

+++ src/main/java/com/example/demo/model/ExchangeRate.java
@@ -2,11 +2,12 @@ package com.example.demo.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Currency;
 import java.util.Map;
 
 @Data
 public class ExchangeRate {
     private Currency base;
-    private Map<Currency, Double> rates;
+    private Map<Currency, BigDecimal> rates;
 }
